refactor: merge readers/ + core/domain/parser/ + core/input/ into input/
Consolidates all file-reading and IBKR-parsing code into a single input/
layer so the boundary between "reading" and "parsing" is no longer a
separate folder distinction.

- src/input/ now owns the full File→InputData pipeline:
  - format readers (readCsv, readPdf, readHtml)
  - IBKR-specific parsers (parseCsvTrades, parseDividends, etc.)
  - validation and assembly (validateInput, parseInput, buildInputData)
  - fileReader.js (browser entry point)

- buildInputData now pre-builds instrumentInfo and csvTradeBasis so
  core services receive fully-assembled InputData and no longer import
  from parser files.

- core/ is now a clean stateless API with two entry points:
  inferPriorPositions(inputData) and calculateTax(inputData, priorPositions).

- core/domain/aliases.js and core/domain/positions.js extracted from
  parseInstruments/parseOpenPositions so HoldingsCalculator stays within
  core/domain/.

- inferPriorPositions signature simplified to accept full InputData.

All 192 tests pass. Build succeeds.
