+++ src/core/services/calculateTax.js
@@ -1,5 +1,3 @@
-import { buildInstrumentInfo } from '../domain/parser/parseInstruments.js'
-import { buildCsvTradeBasis } from '../domain/parser/parseCsvTrades.js'
 import {
   getLocalCurrencyCode,
   getLocalCurrencyLabel,
@@ -31,8 +29,7 @@ export function calculateTax(input, priorPositions = [], { strategy = 'ibkr' } =
 
   const context = buildTaxContext(taxYear)
 
-  const instrumentInfo = buildInstrumentInfo(input.instruments)
-  const csvTradeBasis = buildCsvTradeBasis(input.csvTrades)
+  const { instrumentInfo, csvTradeBasis } = input
 
   // --- TRADES ---
   const tradeCalc = new TradeCalculator({ instrumentInfo, csvTradeBasis, context, strategy })
