+++ src/input/__tests__/parseInterest.test.js
