+++ src/input/parseInput.js
@@ -13,17 +13,17 @@
  *   buildInputData(csvRows, trades)     → InputData
  */
 
-import { parseCSV }                                        from '../../readers/readCsv.js'
-import { PdfTableExtractor }                                from '../domain/parser/PdfTableExtractor.js'
-import { parseTradePdf }                                   from '../domain/parser/PdfTradeConfirmationParser.js'
-import { parseTradesFromHtml }                             from '../domain/parser/parseTradesHtml.js'
-import { parseStatementInfo }                              from '../domain/parser/parseStatementInfo.js'
-import { parseInstruments }                                from '../domain/parser/parseInstruments.js'
-import { parseDividends, parseWithholdingTax }             from '../domain/parser/parseDividends.js'
-import { parseOpenPositions }                              from '../domain/parser/parseOpenPositions.js'
-import { parseCsvTrades }                                  from '../domain/parser/parseCsvTrades.js'
-import { parseInterest }                                   from '../domain/parser/parseInterest.js'
-import { parseTaxYear }                                    from '../domain/parser/parseTaxYear.js'
+import { parseCSV }                                        from './readCsv.js'
+import { PdfTableExtractor }                                from './PdfTableExtractor.js'
+import { parseTradePdf }                                   from './PdfTradeConfirmationParser.js'
+import { parseTradesFromHtml }                             from './parseTradesHtml.js'
+import { parseStatementInfo }                              from './parseStatementInfo.js'
+import { parseInstruments, buildInstrumentInfo }          from './parseInstruments.js'
+import { parseDividends, parseWithholdingTax }             from './parseDividends.js'
+import { parseOpenPositions }                              from './parseOpenPositions.js'
+import { parseCsvTrades, buildCsvTradeBasis }              from './parseCsvTrades.js'
+import { parseInterest }                                   from './parseInterest.js'
+import { parseTaxYear }                                    from './parseTaxYear.js'
 import {
   validateCsvContent,
   validateHtmlContent,
@@ -35,22 +35,12 @@ import {
 // Activity Statement
 // ---------------------------------------------------------------------------
 
-/**
- * Parse an Activity Statement CSV into normalized rows.
- * @param {string} csvText
- * @returns {string[][]}
- */
 export function parseActivityStatementCsv(csvText) {
   const rows = parseCSV(csvText)
   validateCsvContent(rows)
   return rows
 }
 
-/**
- * Parse an Activity Statement PDF into normalized rows.
- * @param {import('../io/readPdf.js').PdfPage[]} pages
- * @returns {string[][]}
- */
 export function parseActivityStatementPdf(pages) {
   const rows = new PdfTableExtractor().adapt(pages)
   validatePdfContent(rows)
@@ -61,11 +51,6 @@ export function parseActivityStatementPdf(pages) {
 // Trade Confirmation
 // ---------------------------------------------------------------------------
 
-/**
- * Parse a Trade Confirmation HTML document into normalized trade objects.
- * @param {Document} doc - pre-parsed HTML document
- * @returns {object[]}
- */
 export function parseTradeConfirmationHtml(doc) {
   validateHtmlContent(doc)
   const trades = parseTradesFromHtml(doc)
@@ -73,11 +58,6 @@ export function parseTradeConfirmationHtml(doc) {
   return trades
 }
 
-/**
- * Parse a Trade Confirmation PDF into normalized trade objects.
- * @param {import('../io/readPdf.js').PdfPage[]} pages
- * @returns {object[]}
- */
 export function parseTradeConfirmationPdf(pages) {
   const trades = parseTradePdf(pages)
   validateTradeCurrencies(trades)
@@ -88,28 +68,24 @@ export function parseTradeConfirmationPdf(pages) {
 // Assembly
 // ---------------------------------------------------------------------------
 
-/**
- * Assemble normalized InputData from parsed activity rows and trades.
- * This is the canonical InputData schema — identical regardless of file source.
- *
- * @param {string[][]} csvRows
- * @param {object[]} trades
- * @returns {InputData}
- */
 export function buildInputData(csvRows, trades) {
   const { statement, account } = parseStatementInfo(csvRows)
-  const taxYear = parseTaxYear(statement.period)
+  const taxYear      = parseTaxYear(statement.period)
+  const instruments  = parseInstruments(csvRows)
+  const csvTrades    = parseCsvTrades(csvRows)
 
   return {
     statement,
     account,
     taxYear,
-    instruments:    parseInstruments(csvRows),
+    instruments,
     dividends:      parseDividends(csvRows),
     withholdingTax: parseWithholdingTax(csvRows),
     trades,
     openPositions:  parseOpenPositions(csvRows),
-    csvTrades:      parseCsvTrades(csvRows),
+    csvTrades,
     interest:       parseInterest(csvRows),
+    instrumentInfo: buildInstrumentInfo(instruments),
+    csvTradeBasis:  buildCsvTradeBasis(csvTrades),
   }
 }
