+++ src/input/validateInput.js
